Refactor middleware validations in service calls
